requirements.txt
colorama~=0.4.4
fpdf2~=2.5.4
pdoc
pyinstaller